set balance type to u128

+++ core/src/transaction.rs
@@ -23,7 +23,7 @@ pub struct Transaction {
 	/// Gas paid up front for transaction execution.
 	pub gas: u64,
 	/// Transfered value.
-	pub value: u64,
+	pub value: u128,
 	pub sign_data: ([u8;32],[u8;32],[u8;32]),
 	/// Transaction data.
 	pub data: Bytes,
@@ -39,14 +39,14 @@ impl Transaction {
 	pub fn get_nonce(&self) -> u64 {
 		self.nonce
 	}
-	pub fn get_value(&self) -> u64 {
+	pub fn get_value(&self) -> u128 {
 		self.value
 	}
 	pub fn get_sign_data(&self) -> SignatureInfo {
 		SignatureInfo::make(self.sign_data.0,self.sign_data.1,self.sign_data.2)
 	}
 	pub fn new(sender: Address, recipient: Address, nonce: u64, gas_price: u64, gas: u64, 
-		value: u64, data: Bytes) -> Transaction {
+		value: u128, data: Bytes) -> Transaction {
         Transaction {
            sender: sender,
             recipient:recipient,
